Move st_h2o_headers_add_arg_t to configurator.h
